pyproject.toml
[tool.poetry]
name = "economic_indicators"
version = "0.1.0"
description = ""
authors = ["benjaminalabaschin@gmail.com"]

[tool.poetry.dependencies]
python = "3.10.4"
streamlit = "^1.14.1"
ipykernel = "^6.17.1"
jupyter = "^1.0.0"
beautifulsoup4 = "^4.11.1"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
matplotlib = "^3.6.2"
plotly = "^5.11.0"
prettytable = "^3.5.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 88
profile = "black"
skip = ["./.venv", ".direnv", ".env"]

[tool.black]
exclude = '''
(
	/(
		\.vscode
		| \.git
		| \.pytest_cache
		| \.mypy_cache
		| \.venv
		| \.env
		| \.direnv
	)/
)
'''
include = '\.pyo?$'
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
files = [
	"app/**/*.py",
	"tests/**/*.py"
]
follow_imports = "silent"
ignore_missing_imports = true
scripts_are_modules = true

